Make a warnings fatal, and minors fixes

+++ PBS/Bookmark.pm
@@ -2,6 +2,7 @@ package PBS::Bookmark;
 
 # pragma
 use autodie; 
+use warnings FATAL => 'all'; 
 
 # core
 use File::Find;
@@ -9,6 +10,7 @@ use File::Find;
 # cpan 
 use Moose::Role;  
 use namespace::autoclean; 
+use Try::Tiny;
 
 # features 
 use experimental qw(signatures);  
@@ -27,16 +29,11 @@ has 'bookmark', (
         my $bookmark; 
         my %mod_time = (); 
         
-        # required permission 
-        if ( $self->owner eq $ENV{USER} ) { 
-            # recursively find modified time of all OUTCAR in directory 
-            find( sub { $mod_time{$File::Find::name} = -M if /OUTCAR/ }, $self->init );  
-
-            # the latest OUTCAR created 
-            if ( %mod_time ) { 
-                $bookmark = ( sort { $mod_time{$a} <=> $mod_time{$b} } keys %mod_time )[0] =~ s/\/OUTCAR//r; 
-            }
-        }  
+        # silent permission error 
+        try { 
+            find( sub { $mod_time{$File::Find::name} = -M if /OUTCAR/ }, $self->init ); 
+            $bookmark = ( sort { $mod_time{$a} <=> $mod_time{$b} } keys %mod_time )[0] =~ s/\/OUTCAR//r; 
+        }; 
 
         return $bookmark //= '';  
     }, 
@@ -44,18 +41,15 @@ has 'bookmark', (
 
 # <modifiers> 
 after 'info' => sub ( $self ) { 
-    my $prefix = $self->init; 
     if ( $self->bookmark ) { 
-        printf "%-9s> %s\n", ucfirst('bookmark'), $self->bookmark =~ s/$prefix\///r;  
+        printf "%-9s> %s\n", ucfirst('bookmark'), $self->bookmark =~ s/${\$self->init}\///r;  
     } 
 }; 
 
 # <methods>
 # reset job by deleting latest OUTCAR  
 sub reset ( $self ) { 
-    if ( $self->bookmark and $self->prompt('reset') ) { 
-        unlink join '/', $self->bookmark, 'OUTCAR'; 
-    }
+    if ( $self->bookmark and $self->prompt('reset') ) { unlink join '/', $self->bookmark, 'OUTCAR' }
 } 
 
 1;
